refactor(parseFeed): 按 SourceKind 拆分为 parseFeed/ 子模块
- shared: 节点取值/日期解析/Article 构造等共享工具
- generic: RSS/Atom/RDF/JSON Feed 通用路径
- netease/zhihu/flightSites/sites/finance/wechat: 站点定制解析器
- dateEnrichment: 晚点/甲子光年/Paul Graham 详情页日期补全
- lib/parseFeed.ts 保留为入口(PARSERS 分派 + 公开 API 再导出), 外部导入路径不变
